backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.3
django-redis==5.4.0
boto3==1.34.69
Pillow==10.2.0
python-decouple==3.8
gunicorn==21.2.0
sentry-sdk==1.43.0
whitenoise==6.6.0
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.1
numpy==1.26.4
openai==1.14.3
tiktoken==0.6.0
stripe==8.8.0
django-storages==1.14.2
django-extensions==3.2.3
djangorestframework-camel-case==1.4.2
drf-spectacular==0.27.1
django-audit-log==0.7.0
requests==2.31.0
python-magic==0.4.27
chardet==5.2.0
phonenumbers==8.13.31
